Today I am further advanced in working with data than ever before.
At the moment, I have finished writing the logic
for assembling a structured date
(By date, I mean a large object, with userData, list, word, etc.)

The data comes from the DB and is stored in fragments.
This will allow them to be updated in small pieces,
sending and receiving small amounts of data

To make them easier to use,
they are collected into an object of objects.
This happens every time you reload

If the data is in loacalstorage, it is taken from there.
If not, then with mongoDB.

The main thing is that I finally + -
figured out promises and “data processing sequence logic”

still need to do:
* adding, deleting data
 * gaming (which includes swipe mechanics), + recording session results.
 * scatter data update handlers wherever necessary
 (add bonus restrictions on the update frequency
 in some cases and an emergency recording when closing the page)